+++ my_run_backend.bat
@@ -1,0 +1,3 @@
+@echo off
+E:\AItools\GPT-SoVITS-beta\runtime\python.exe .\inference_backend.py
+pause
